requirements.txt
pandas==1.5.3
numpy==1.23.3
matplotlib==3.6.2
scikit-learn==1.1.2
xgboost==1.6.2  # optional, required for XGB models
tensorflow==2.11.0  # optional, required for LSTM models
scipy==1.9.3

# Development / optional tools
jupyter
pyyaml==6.0
mlflow==2.2.2  # optional, experiment tracking
streamlit==1.18.0  # dashboard
altair==4.2.2
optuna==3.1.0  # optional, for hyperparameter optimization
